pyproject.toml
[project]
name = "ai-course10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen>=0.9.3",
    "crewai>=0.130.0",
    "docling>=2.37.0",
    "duckduckgo-search>=8.0.4",
    "google-generativeai>=0.8.5",
    "gradio>=5.34.1",
    "ipykernel>=6.29.5",
    "lancedb>=0.23.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.24",
    "mcp[cli]>=1.9.4",
    "openai>=1.88.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.45.1",
    "tiktoken>=0.9.0",
]
